Add `Hint` for fast recovery.
`Hint` is a special kind of log file which only save the indexes
of log entries that refer to the real log file. When recovering
RaftEngine, needs to restore memtable which maintains all log
indexes. Previously, RaftEngine had to scan all of the log files
which is slow and reading actuall data is unneeded.

This implemenation constructs and flushes hint file in a background
thread.

TODO:
- [x] Implement hint worker and tasks.
- [x] Implement hint file format.
- [ ] Save to & restore from file.
- [ ] Unittests.
- [ ] Benchmark.

Signed-off-by: MrCroxx <mrcroxx@outlook.com>