rollback to a consistent state when fsync error
invariant: non-active log files are always synced and safe.

Introduce `pre_write` and `post_write` for `PipeLog` to acquire
or release a `WriteContext`, which can be used to track the sync
state.

If `fsync` raises an error, log will rollback to last synced state,
and set errors for the writers in the current write group. But if
there are bytes rollbacked but in writers in the previous write
groups, RaftEngine will panic, because that means some data that
the user thinks is safe but lost actually.

`new_log_file` is devided into 2 phase. First sync old log file,
then rotate new log file. The second phase is safe to retry if
`sync_dir` fails, because the new log will overwrite the failed
one.

Signed-off-by: MrCroxx <mrcroxx@outlook.com>