improve exception safety of FilePipeLog and overhaul failpoint tests (#147)
We should guarantee the in-memory states of objects are still consistent after exception (error or panic). i.e. We should be able to operate on `Engine` even after catching a panic from it. Reorder some the execution to achieve that, and add test cases for verification.

Some other considerations during overhaul:
- Avoid `unwrap`-ing inside a `catch_unwind` block
- Clean up `ConcurrentWriteContext`

Signed-off-by: tabokie <xy.tao@outlook.com>