Redesign the implementation on `recover_queue` for better performance and compatibility.
Considering the compatibility to `VERSION` and good performance on `recovery`, the implementation on `recover_queue` is modified here.
And compared with benchmark on `master`, the result of benchmark on current PR shows no performance regression.
Configuration of my local environment for benchmark was listed as following:
> * OS: MacOS - Monterey 12.1
> * Mem: 16GB
> * CPU: Apple M1 Pro
> * DISK: NVME SSD - APPLE SSD AP0512R

By `cargo bench -- --bench benches`, the performance between `master` branch and current PR could be reviewed with the following details:
>* `Master`: 108.99ms on `default` workload, 122.41ms on `compressed` workload, 994ms on `mini-batch(1kb)` workload, 2.2368s on `10GB` workload.
>* current PR: 111.63ms on `default` workload, 116.35ms on `compressed` workload, 921.98ms on `mini-batch(1kb)` workload, 2.0905s on `10GB` workload.

All the results are generated by `bench` on a same dataset, and the minor fluctuation on performance resulted by other processes.

Signed-off-by: Lucasliang <nkcs_lykx@hotmail.com>