fix data loss of partial rewrite with atomic group (#290)
A second attempt to fix #288.

This PR introduces the concept of "atomic group". It is then used by rewrite-rewrite operation to make sure the rewrite of each region is perceived as an atomic operation.

A group of writes is made atomic by each carrying a special marker. During recovery, log batch with the marker will be stashed until all parts of the group are found. Caveats for this approach is commented near `AtomicGroupBuilder`.

Also fixed a bug that a partial rewrite-rewrite (due to batch being split) is not applied correctly.

Signed-off-by: tabokie <xy.tao@outlook.com>